+++ app/models/protonym.rb
@@ -257,26 +257,6 @@ class Protonym < TaxonName
     end
   end
 
-  # @return [String, nil]
-  # TODO: not true, *has* parens?!
-  #   a string, without parenthesis, that includes author and year
-  def get_author_and_year
-    # times_called
-    case rank_class.try(:nomenclatural_code)
-      when :iczn
-        ay = iczn_author_and_year
-      when :icvcn
-        ay = icn_author_and_year
-      when :icnp
-        ay = icn_author_and_year
-      when :icn
-        ay = icn_author_and_year
-      else
-        ay = ([author_string] + [year_integer]).compact.join(' ')
-    end
-    ay.blank? ? nil : ay
-  end
-
   # This method is currently only used for setting cached_primary_homonym
   # @return [nil, false, String]
   #   !! Why both?